Split into two directories
