+++ src/core/build_content.js
