+++ src/core/commit_flags.js
