+++ src/core/guru_md_block.js
