+++ src/core/handle_card.js
