+++ src/core/hast_util.js
