+++ src/core/input_builder.js
