+++ src/core/version.cjs
