+++ src/gh_action/index.js
@@ -2,15 +2,15 @@ import { performance } from 'perf_hooks';
 import core from '@actions/core';
 import c from 'ansi-colors';
 
-import getInputs from './inputs.js';
+import getInputs from './inputs.js.js';
 import ghLogger from './gh_logger.js';
 import createClient from './api_client.js';
 import createFetch from './fetch.js';
-import version from './version.cjs';
-import { isRepoPublic } from './util.js';
+import version from './version.cjs.js';
+import { isRepoPublic } from './util.js.js';
 import commitCardsFile from './commit_cards_file.js';
 import getChangedFilesBase from './file_changes.js';
-import action from './action.js';
+import action from './core/action.js.js';
 
 /**
  * This is the main entry point for the GitHub action. It's SOLE purpose should be to set up dependencies for
