+++ test/action.test.js
@@ -1,12 +1,12 @@
-import runAction from '../src/action.js';
+import runAction from '../src/core/action.js';
 import createClient from './support/api_client.js';
 import { resource } from './support/util.js';
 import fs from 'fs';
-import { readFile, writeFile } from '../src/fs_util.js';
+import { readFile, writeFile } from '../src/core/fs_util.js';
 import arrayLogger from './support/array_logger.js';
 import nullLogger from './support/null_logger.js';
 import nullColorizer from './support/null_colorizer.js';
-import { InvalidGitObjectError } from '../src/error.js';
+import { InvalidGitObjectError } from '../src/core/error.js';
 
 beforeEach(async() => {
     if(fs.existsSync('test/env')) {
