+++ test/support/util.js
@@ -1,5 +1,5 @@
 import path from 'path';
-import { readFile } from '../../src/fs_util.js';
+import { readFile } from '../../src/core/fs_util.js';
 
 export async function resource(filePath) {
     return await readFile(path.join('test/resources', filePath));
